[OM-64385]: RB:42566 Apply scaling factor consistently during market conversion
** Problem **

At Bank of America, we see an IllegalArgumentException being thrown during market analysis when we attempt to set a commodity's quantity to a very slightly negative number (approximately 1e-4).

After debugging for some time, we identified that the cause was a resize action on a Container's VCPU sold from a capacity of 1491.3689 to 0. When propagating the impact of this resize to dependent commodities, the used value on VCPU_LIMIT_QUOTA bought from the ContainerPod has a value of 1491.369. We subtract off this 1491.369 from the quantity sold by the ContainerPod, but the used value on the ContainerPod's sold commodity is only 1491.3689, and so the result of the subtraction is a negative number resulting in the IllegalArgument.

When looking in the topology diagnostics, we see that both the VCPU_LIMIT_QUOTA bought by the Container and sold by the pod are equal (a value of 540.0). However, a scalingFactor of 2.7617943704573737 is being applied and after the multiplication by scaling factor, we see this difference.

Guy has already hardened market code to prevent this sort of exception from completely halting market analysis, but there are still scenarios where this sort of floating point precision problem can lead to incorrect decisions (particularly around the capacityUpperBound and capacityLowerBound). We should fix this to try to be as consistent as possible in our handling of the scalingFactor to ensure correct market behavior.

** Solution **

Because most commodity quantities in TopologyEntityDTOs are doubles but are floats in the TraderTO's we create for the market, we have to incur some floating point precision loss during the conversion. We just need to ensure that we incur the loss in the same way for all quantities multiplied by the scaling factor, otherwise certain numerical comparisons we do (whether with 0 when setting quantities or with the commodity upper/lower bounds) may fail.

This fix updates commodity conversion logic so that we apply scaling factor logic by incurring floating point precision loss the same way for relevant fields on commodities bought and sold.

** Testing Done **
* New unit test to ensure scaling factor is applied consistently during market conversion.

+++ com.vmturbo.market.component/src/main/java/com/vmturbo/market/topology/conversions/CommodityConverter.java
@@ -145,7 +145,7 @@ public class CommodityConverter {
                         : null);
         // if this commodity has a scaling factor set, then scale up the
         // USED and CAPACITY by scalingFactor for use in the new CommoditySoldTO
-        final float scalingFactor = (float)topologyCommSold.getScalingFactor();
+        final double scalingFactor = topologyCommSold.getScalingFactor();
         if (topologyCommSold.hasScalingFactor() && logger.isDebugEnabled()) {
             logger.debug("Scaling comm {}, factor {}, for topology entity {},"
                             + " prev used {}, new {}, prev capacity {}, new {}",
@@ -239,21 +239,21 @@ public class CommodityConverter {
         final CommodityDTOs.CommoditySoldSettingsTO.Builder economyCommSoldSettings =
                 CommodityDTOs.CommoditySoldSettingsTO.newBuilder()
                         .setResizable(resizable)
-                        .setCapacityIncrement(topologyCommSold.getCapacityIncrement() * scalingFactor)
+                        .setCapacityIncrement((float)(topologyCommSold.getCapacityIncrement() * scalingFactor))
                         .setUtilizationUpperBound(utilizationUpperBound)
                         .setPriceFunction(priceFunction(topologyCommSold.getCommodityType(),
-                                scale, dto))
+                            scale, dto))
                         .setUpdateFunction(updateFunction(topologyCommSold, commodityTypeAllocator));
 
         // Set thresholds for the commodity sold (min/Max of VCPU/VMem for on-prem VMs).
         if (topologyCommSold.hasThresholds()) {
             final Thresholds threshold = topologyCommSold.getThresholds();
             if (threshold.hasMax()) {
-                final float maxThreshold = Double.valueOf(threshold.getMax()).floatValue() * scalingFactor;
+                final float maxThreshold = (float)((float)threshold.getMax() * scalingFactor);
                 economyCommSoldSettings.setCapacityUpperBound(maxThreshold);
             }
             if (threshold.hasMin()) {
-                final float minThreshold = Double.valueOf(threshold.getMin()).floatValue() * scalingFactor;
+                final float minThreshold = (float)((float)threshold.getMin() * scalingFactor);
                 economyCommSoldSettings.setCapacityLowerBound(minThreshold);
             }
 
@@ -335,7 +335,7 @@ public class CommodityConverter {
                         topologyCommSold.getHistoricalUsed().getPercentile(),
                         topologyCommSold.getCommodityType().getType(),
                         dto.getDisplayName());
-                soldCommBuilder.setHistoricalQuantity((float)(topologyCommSold.getCapacity()
+                soldCommBuilder.setHistoricalQuantity((float)((float)topologyCommSold.getCapacity()
                         * topologyCommSold.getScalingFactor()
                         * topologyCommSold.getHistoricalUsed().getPercentile()));
             }
